+++ extensions/bitwarden/src/components/UnlockForm.tsx
@@ -60,7 +60,8 @@ const UnlockForm = ({ pendingAction = Promise.resolve() }: UnlockFormProps) => {
       toast.title = "Unlocking vault...";
       const { error: unlockError } = await bitwarden.unlock(password);
       if (unlockError) {
-        if (unlockError instanceof InvalidSessionTokenError) {
+        if (unlockError instanceof InvalidSessionTokenError && !args?.retryInvalidSessionToken) {
+          await bitwarden.logout({ reason: "Invalid session token" });
           return onSubmit({ retryInvalidSessionToken: true });
         }
         return handleUnlockError(unlockError, {